Rename some stuff for better readability
