+++ jump-core/src/main/java/com/bitdecay/jump/BitBody.java
@@ -1,7 +1,7 @@
 package com.bitdecay.jump;
 
 import com.bitdecay.jump.collision.ContactListener;
-import com.bitdecay.jump.controller.BitBodyController;
+import com.bitdecay.jump.control.BitBodyController;
 import com.bitdecay.jump.exception.BitBodySerializeException;
 import com.bitdecay.jump.geom.BitPoint;
 import com.bitdecay.jump.geom.BitRectangle;
