+++ jump-core/src/main/java/com/bitdecay/jump/control/PathedBodyController.java
@@ -1,4 +1,4 @@
-package com.bitdecay.jump.controller;
+package com.bitdecay.jump.control;
 
 import com.bitdecay.jump.BitBody;
 import com.bitdecay.jump.collision.BitWorld;
