+++ jump-core/src/main/java/com/bitdecay/jump/control/state/JumperBodyControlState.java
@@ -1,0 +1,12 @@
+package com.bitdecay.jump.control.state;
+
+import com.bitdecay.jump.JumperBody;
+import com.bitdecay.jump.control.ControlMap;
+
+/**
+ * Created by Monday on 11/3/2015.
+ */
+public interface JumperBodyControlState {
+    void stateEntered(JumperBody body, ControlMap controls);
+    JumperBodyControlState update(float delta, JumperBody body, ControlMap controls);
+}
