+++ jump-core/src/main/java/com/bitdecay/jump/control/state/SidewaysControlState.java
@@ -1,15 +1,15 @@
-package com.bitdecay.jump.controller.state;
+package com.bitdecay.jump.control.state;
 
 import com.bitdecay.jump.BitBody;
 import com.bitdecay.jump.Facing;
-import com.bitdecay.jump.controller.ControlMap;
-import com.bitdecay.jump.controller.PlayerAction;
+import com.bitdecay.jump.control.ControlMap;
+import com.bitdecay.jump.control.PlayerAction;
 import com.bitdecay.jump.geom.BitPoint;
 
 /**
  * Created by Monday on 11/4/2015.
  */
-public abstract class SidewaysState implements JumperControlState {
+public abstract class SidewaysControlState implements JumperBodyControlState {
     protected void handleLeftRight(float delta, BitBody body, ControlMap controls, BitPoint accel, BitPoint decel) {
         boolean requestLeft = controls.isPressed(PlayerAction.LEFT);
         boolean requestRight = controls.isPressed(PlayerAction.RIGHT);
